[tilck] Add a realpath() impl. which doesn't use /proc
Tilck doesn't currently support /proc but it uses a libmusl-based toolchain
for building user applications. Unfortunately, libmusl's implementation of
realpath() requires the symlink /proc/self/fd/<n> to exist and it has no
fallback for proc-less systems. The "right" fix might be to patch libmusl,
but it's in the philosophy of Tilck's project to use mainly pre-built GCC
toolchains downloaded from https://toolchains.bootlin.com. Therefore, in order
to make realpath() work, Tilck has either to switch to a different kind of
toolchain (e.g. uclibc-ng) or to add at least a partial support for /proc in
the Tilck kernel itself. In both the cases, that would require spending a
considerable amount of effort. If micropython is the only program that will
benefit from that, it's just not worth doing at this time. Therefore, the
most convenient "fix" currently is to just patch micropython and make it use
a custom version of realpath(), not depending on /proc, which is exactly what
this patch does.